requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-pptx>=0.6.23
openai>=1.12.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
